requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
numpy==1.23.5
tensorflow==2.12.0
keras==2.12.0
Pillow==9.5.0
gunicorn==22.0.0